ServiceDirectory: mark flags as required (#5)
`region` and `project` are now required flags. This doesn't change much the
behavior of the program, as an error was being thrown when they were
empty even before this change.

Labelling them as required will now make the framework always
display the usage when they are empty.
Also, this frees us from having to validate them explicitly.

Signed-off-by: Elis Lulja <elulja@cisco.com>