Move credentials flag to servicedirectory command (#5)
In an effort to make the CLI more descriptive, this commit removes the
`--credentials` flag from the root command and moves it to the
`servicedirectory` command with new name `--service-account`.
With this change the `--service-account` flag now becomes a *local* flag,
not a global one as was `--credentials` before.

This is because Google Cloud documentation always talks about
*Service Account* rather than credentials, so this change should help
anyone understand the command at first glimpse.

Readme will be updated in next commit.

Signed-off-by: Elis Lulja <elulja@cisco.com>