etcd: process create events (#26)
In this commit create events are processed according to their type.
Only new events concerning endpoints are processed, as those are the
ones that are really sent externally.
Parsing is done through a function designed to be re-used as much
as possible for other event types.
Unit tests are included.

Signed-off-by: Elis Lulja <elulja@cisco.com>